requirements.txt
Django == 5.0.2
psycopg[binary]
django-bootstrap5
django-filter
django-debug-toolbar
python-dateutil
